FocusableTextFieldAdapter: clean password field after incorrect login data fix. Make SwiftUI UIKit connection with Coordinator
